comparisons done (i think)

+++ src/main/java/logic/SortResult.java
@@ -3,9 +3,9 @@ package logic;
 public class SortResult {
     public long comparisons;
     public long interchanges;
-    public long averageRunningTime;
-    public long minRunningTime;
-    public long maxRunningTime;
+    public double averageRunningTime;
+    public double minRunningTime;
+    public double maxRunningTime;
     public String algorithmName;
     public int arraySize;
     public String arrayType;
